Pack Sidebar in its own module
