+++ src/Search.tsx
@@ -1,38 +1,25 @@
-import { useCallback, useRef } from "react";
 import {
   Box,
+  Card,
+  CardBody,
+  Center,
   Flex,
-  useDisclosure,
   Grid,
   GridItem,
   Heading,
-  Center,
-  Card,
-  CardBody,
-  Drawer,
-  DrawerOverlay,
-  DrawerContent,
-  DrawerHeader,
-  InputGroup,
-  Input,
-  IconButton,
-  DrawerCloseButton,
-  DrawerBody,
-  useColorModeValue,
-  Text,
-  keyframes,
+  useDisclosure,
 } from "@chakra-ui/react";
-import { type LoaderFunctionArgs, redirect, useLoaderData, Form } from "react-router-dom";
-import { TbListSearch, TbSearch } from "react-icons/tb";
+import { useCallback, useRef } from "react";
+import { TbListSearch } from "react-icons/tb";
+import { redirect, useLoaderData, type LoaderFunctionArgs } from "react-router-dom";
 
 import Header from "./components/Header";
-import Sidebar from "./components/Sidebar";
-import db, { ChatCraftMessageTable } from "./lib/db";
 import Message from "./components/Message";
-import { ChatCraftMessage } from "./lib/ChatCraftMessage";
 import OptionsButton from "./components/OptionsButton";
+import Sidebar from "./components/Sidebar/";
 import { useSettings } from "./hooks/use-settings";
-import useMobileBreakpoint from "./hooks/use-mobile-breakpoint";
+import { ChatCraftMessage } from "./lib/ChatCraftMessage";
+import db, { ChatCraftMessageTable } from "./lib/db";
 
 export async function loader({ request }: LoaderFunctionArgs) {
   const url = new URL(request.url);
@@ -86,32 +73,6 @@ export default function Search() {
     setSettings({ ...settings, sidebarVisible: newValue });
   }, [isSidebarVisible, settings, setSettings, toggleSidebarVisible]);
 
-  const isMobile = useMobileBreakpoint();
-  const sidebarColor = useColorModeValue("blue.600", "blue.200");
-
-  const sidebarOpenAnimationKeyframes = keyframes`
-    from {
-      opacity: 0;
-    }
-
-    to {
-      opacity: 1;
-    }
-  `;
-
-  const sidebarCloseAnimationKeyframes = keyframes`
-    from {
-      transform: scaleX(1);
-    }
-
-    to {
-      transform: scaleX(0);
-    }
-  `;
-
-  const sidebarOpenAnimation = `${sidebarOpenAnimationKeyframes} 500ms ease-in-out forwards`;
-  const sidebarCloseAnimation = `${sidebarCloseAnimationKeyframes} 100ms ease-in-out forwards`;
-
   return (
     <Grid
       w="100%"
@@ -134,55 +95,11 @@ export default function Search() {
       </GridItem>
 
       <GridItem rowSpan={3} overflowY="auto">
-        {isMobile ? (
-          <Drawer isOpen={isSidebarVisible} onClose={handleToggleSidebarVisible} placement="left">
-            <DrawerOverlay />
-            <DrawerContent>
-              <DrawerHeader mt={2} p={2}>
-                <Text
-                  position={"relative"}
-                  top={-1}
-                  ml={2}
-                  mb={2}
-                  fontSize="lg"
-                  fontWeight="bold"
-                  color={sidebarColor}
-                >
-                  &lt;ChatCraft /&gt;
-                </Text>
-                <Form action="/s" method="get" onSubmit={handleToggleSidebarVisible}>
-                  <InputGroup size="sm" variant="outline">
-                    <Input
-                      type="search"
-                      defaultValue={searchText}
-                      name="q"
-                      isRequired
-                      placeholder="Search chat history"
-                    />
-                    <IconButton
-                      aria-label="Search"
-                      variant="ghost"
-                      icon={<TbSearch />}
-                      type="submit"
-                    />
-                  </InputGroup>
-                </Form>
-                <DrawerCloseButton />
-              </DrawerHeader>
-
-              <DrawerBody m={0} p={0}>
-                <Sidebar />
-              </DrawerBody>
-            </DrawerContent>
-          </Drawer>
-        ) : (
-          <Box
-            transformOrigin={"left"}
-            animation={isSidebarVisible ? sidebarOpenAnimation : sidebarCloseAnimation}
-          >
-            <Sidebar />
-          </Box>
-        )}
+        <Sidebar
+          searchText={searchText}
+          isSidebarVisible={isSidebarVisible}
+          handleToggleSidebarVisible={handleToggleSidebarVisible}
+        ></Sidebar>
       </GridItem>
 
       <GridItem overflowY="auto" ref={messageListRef} pos="relative">
